Extract noopHook into its own file in pkg/hook
Move the noopHook implementation and NewNoopHook constructor from hook.go
into a dedicated noop_impl.go file to keep hook.go focused on the core
publish/subscribe mechanics.
